backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
sentence-transformers
faiss-cpu
jieba
scikit-learn
numpy
pandas
openpyxl
PyPDF2
python-docx
markdown
requests
Werkzeug
pytest
pytest-flask

